requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
dj-places==4.0.0
Django==3.1.7
django-environ==0.4.5
django-tawkto==0.3
django-bootstrap-v5==1.0.1
django-braces==1.14.0
django-crispy-forms==1.11.2
django-multiselectfield==0.1.12
django-sendgrid-v5==0.9.1
django-oauth-toolkit==1.5.0
djangorestframework==3.12.4
drf-social-oauth2==1.0.9
drf-yasg==1.20.0
flake8==3.9.0
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jwcrypto==0.8
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.1.2
pycparser==2.20
PyJWT
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.2
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core
soupsieve==2.2.1
sqlparse==0.4.1
stripe==2.56.0
twilio
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.2.0
uwsgi>=2.0.18,<2.1.0
wheel
psycopg2>=2.7.5,<2.8.0
